rename fork tool folder
