+++ src/tools/index.ts
@@ -1,13 +1,13 @@
 // AST-grep tools
 export { ast_grep_replace, ast_grep_search } from './ast-grep';
 export { createCouncilTool } from './council';
-export type { ForkCommandManager } from './handoff';
+export type { ForkCommandManager } from './fork';
 export {
   createForkCommandManager,
   createForkSessionTool,
   createForkState,
   createReadSessionTool,
-} from './handoff';
+} from './fork';
 export type { PresetManager } from './preset-manager';
 export { createPresetManager } from './preset-manager';
 export { createWebfetchTool } from './smartfetch';
